internal/ui: begin C rewrite
